Styling changes

+++ app/forms/update_funds_form.py
@@ -1,6 +1,6 @@
 from flask_wtf import FlaskForm
-from wtforms import IntegerField
+from wtforms import DecimalField
 from wtforms.validators import DataRequired
 
 class UpdateFundsForm(FlaskForm):
-    funds = IntegerField('funds', validators=[DataRequired()])
+    funds = DecimalField('funds', validators=[DataRequired()])
